sdk/python/requirements.txt
# Core dependencies
httpx>=0.25.0
websockets>=11.0.0
pydantic>=2.0.0
PyJWT>=2.8.0
python-dateutil>=2.8.0

# Optional dependencies for plugins
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Template engine for code generation
jinja2>=3.1.0
pyyaml>=6.0

# Caching
cachetools>=5.3.0

# Async utilities
asyncio-throttle>=1.0.0